As part of this commit I have updated the Travis config so that it uses the CloudFoundry deployment provider to deploy the applications to the various environments. What’s more, I have enabled rolling deployments so there is no downtime when we do deployments. Importantly, this means that if the new code won’t deploy, we will not take down the existing code deployment (which is what currently happens).
Now that we don’t use terraform for deployments, this code can (or should) be moved to a separate repository that looks after the CAS infrastructure.

Note, I would have preferred to have written the deployment script once and then use some conditional logic to determine which environment to deploy to. However, as not all deployments are linked to static branches, I have had to split the deployment parts up to manage the releases to Sandbox (2 and 3).